Release 1.100.0 .
This also updates Gradle to 6.9, fixes a few versions in the Gradle dependencies, replaces jcenter() usage, and makes the README.md more accurate.
Again, such huge thanks to @raeleus ! I was convinced this update was going to be a huge pain for me to do, but the Zebra Champion did all the hard work!