Bring back some GWT logging code as an option.
The code was removed in https://github.com/libgdx/libgdx/pull/5648 , but I haven't been able to get the fallback mentioned in that PR to work at all. So, I made it a boolean option in GwtApplicationConfiguration if you the HTML app to create the console only when requested. The default is false, and everything behaves the same as it currently does unless you set canCreateLog to true. There's also some comment cleanup in GwtApplicationConfiguration, like a missing asterisk that made one set of what should have been JavaDocs into regular comments.